Update WeatherForecast to contain multiple addresses
The WeatherForecast model, service, and controller were updated to allow for multiple addresses per weather forecast. In the model, the Address property was changed to a list of Addresses. The service was modified to generate multiple random addresses, and the controller now includes a new API endpoint for getting addresses based on the forecast id.